+++ devicemanagement/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 - [CI] Adapt pipeline and conanfile.py to gitlab.com
+- [SW-17936] Fixed Temperature threshold test case to restore threshold value at the end
 ### Security
 
 ## [0.11.0] - 2023-06-02